+++ src/audit/artipacked.rs
@@ -74,14 +74,18 @@ impl Audit for Artipacked {
                 };
 
                 if uses.matches("actions/checkout") {
-                    match with.get("persist-credentials") {
-                        Some(EnvValue::Boolean(false)) => continue,
-                        Some(EnvValue::Boolean(true)) => {
+                    match with
+                        .get("persist-credentials")
+                        .map(|v| v.to_string())
+                        .as_deref()
+                    {
+                        Some("false") => continue,
+                        Some("true") => {
                             // If a user explicitly sets `persist-credentials: true`,
                             // they probably mean it. Only report if in auditor mode.
                             vulnerable_checkouts.push((step, Persona::Auditor))
                         }
-                        // TODO: handle expressions and literal strings here.
+                        // TODO: handle expressions here.
                         // persist-credentials is true by default.
                         _ => vulnerable_checkouts.push((step, Persona::default())),
                     }